Fix linting errors and improve code quality
- Fixed TypeScript errors in monitoring.ts by replacing 'any' types with proper type assertions
- Fixed Function type usage with proper function signature types
- Moved metadata and viewport exports to separate files to fix React fast refresh warnings
- Removed unused Metadata import from layout.tsx
- All linting errors resolved, tests passing, type checking clean
